Move sourcecode/editorextensions --> sourcecode/extensions
