+++ spyder/widgets/sourcecode/extensions/closequotes.py
