+++ spyder/widgets/sourcecode/extensions/manager.py
