+++ spyder/widgets/sourcecode/extensions/tests/__init__.py
